// ui/src/proto/appmodel.ts
export interface ApplicationProcess {
  name: string;
  connections: ApplicationConnection[];
  /** Arguments passed to this process. */
  arguments: string;
  /** Hash to identify this process in the process tree. */
  hash: string;
  /** Child processes of this process. */
  children: ApplicationProcess[];
}

export interface ApplicationConnection {
  destinationName: string;
  destinationPort: number;
  bytesSent: number;
  bytesReceived: number;
}
// ...
export const ApplicationProcess: MessageFns<ApplicationProcess> = {
  encode(message: ApplicationProcess, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.name !== "") {
      writer.uint32(10).string(message.name);
    }
    for (const v of message.connections) {
      ApplicationConnection.encode(v!, writer.uint32(18).fork()).join();
    }
    if (message.arguments !== "") {
      writer.uint32(26).string(message.arguments);
    }
    if (message.hash !== "") {
      writer.uint32(34).string(message.hash);
    }
    for (const v of message.children) {
      ApplicationProcess.encode(v!, writer.uint32(42).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ApplicationProcess {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseApplicationProcess();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.name = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.connections.push(ApplicationConnection.decode(reader, reader.uint32()));
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.arguments = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.hash = reader.string();
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.children.push(ApplicationProcess.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ApplicationProcess {
    return {
      name: isSet(object.name) ? globalThis.String(object.name) : "",
      connections: globalThis.Array.isArray(object?.connections)
        ? object.connections.map((e: any) => ApplicationConnection.fromJSON(e))
        : [],
      arguments: isSet(object.arguments) ? globalThis.String(object.arguments) : "",
      hash: isSet(object.hash) ? globalThis.String(object.hash) : "",
      children: globalThis.Array.isArray(object?.children)
        ? object.children.map((e: any) => ApplicationProcess.fromJSON(e))
        : [],
    };
  },

  toJSON(message: ApplicationProcess): unknown {
    const obj: any = {};
    if (message.name !== "") {
      obj.name = message.name;
    }
    if (message.connections?.length) {
      obj.connections = message.connections.map((e) => ApplicationConnection.toJSON(e));
    }
    if (message.arguments !== "") {
      obj.arguments = message.arguments;
    }
    if (message.hash !== "") {
      obj.hash = message.hash;
    }
    if (message.children?.length) {
      obj.children = message.children.map((e) => ApplicationProcess.toJSON(e));
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<ApplicationProcess>, I>>(base?: I): ApplicationProcess {
    return ApplicationProcess.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ApplicationProcess>, I>>(object: I): ApplicationProcess {
    const message = createBaseApplicationProcess();
    message.name = object.name ?? "";
    message.connections = object.connections?.map((e) => ApplicationConnection.fromPartial(e)) || [];
    message.arguments = object.arguments ?? "";
    message.hash = object.hash ?? "";
    message.children = object.children?.map((e) => ApplicationProcess.fromPartial(e)) || [];
    return message;
  },
};

function createBaseApplicationConnection(): ApplicationConnection {
  return { destinationName: "", destinationPort: 0, bytesSent: 0, bytesReceived: 0 };
}

export const ApplicationConnection: MessageFns<ApplicationConnection> = {
  encode(message: ApplicationConnection, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.destinationName !== "") {
      writer.uint32(10).string(message.destinationName);
    }
    if (message.destinationPort !== 0) {
      writer.uint32(16).uint64(message.destinationPort);
    }
    if (message.bytesSent !== 0) {
      writer.uint32(24).uint64(message.bytesSent);
    }
    if (message.bytesReceived !== 0) {
      writer.uint32(32).uint64(message.bytesReceived);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ApplicationConnection {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseApplicationConnection();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.destinationName = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.destinationPort = longToNumber(reader.uint64());
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.bytesSent = longToNumber(reader.uint64());
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.bytesReceived = longToNumber(reader.uint64());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ApplicationConnection {
    return {
      destinationName: isSet(object.destinationName) ? globalThis.String(object.destinationName) : "",
      destinationPort: isSet(object.destinationPort) ? globalThis.Number(object.destinationPort) : 0,
      bytesSent: isSet(object.bytesSent) ? globalThis.Number(object.bytesSent) : 0,
      bytesReceived: isSet(object.bytesReceived) ? globalThis.Number(object.bytesReceived) : 0,
    };
  },

  toJSON(message: ApplicationConnection): unknown {
    const obj: any = {};
    if (message.destinationName !== "") {
      obj.destinationName = message.destinationName;
    }
    if (message.destinationPort !== 0) {
      obj.destinationPort = Math.round(message.destinationPort);
    }
    if (message.bytesSent !== 0) {
      obj.bytesSent = Math.round(message.bytesSent);
    }
    if (message.bytesReceived !== 0) {
      obj.bytesReceived = Math.round(message.bytesReceived);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<ApplicationConnection>, I>>(base?: I): ApplicationConnection {
    return ApplicationConnection.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ApplicationConnection>, I>>(object: I): ApplicationConnection {
    const message = createBaseApplicationConnection();
    message.destinationName = object.destinationName ?? "";
    message.destinationPort = object.destinationPort ?? 0;
    message.bytesSent = object.bytesSent ?? 0;
    message.bytesReceived = object.bytesReceived ?? 0;
    return message;
  },
};

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}